ran opcode_generator with correct number of 'imm' field bits for I-format

+++ src/instructions.h
@@ -36,8 +36,8 @@ typedef struct {
     uint8_t func3;
     // <TotalBits=5>
     REGISTER rs1;
-    // <TotalBits=8>
-    int8_t imm;
+    // <TotalBits=13>
+    int16_t imm;
 } I_INSTRUCTION;
 I_INSTRUCTION as_i_instruction(uint32_t instruction);
